fission-a2a-env/pyproject.toml
[project]
name = "fission-a2a-env"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.10",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary,pool]>=3.2.9",
    "pydantic>=2.11.7",
    "redis>=6.2.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-dotenv>=0.5.2",
]
